+++ python/acl_anthology/venues.py
@@ -62,7 +62,10 @@ class Venue:
     )
     oldstyle_letter: Optional[str] = field(default=None)
     url: Optional[str] = field(default=None)
-    type: Optional[str] = field(default=None)  # TODO: should we deprecate this?
+    # TODO: Should we reconsider 'type'? Currently used to designate journals
+    # at the venue level; but journals are also marked on the individual
+    # volumes.
+    type: Optional[str] = field(default=None)
 
     @classmethod
     def load_from_yaml(cls, path: PathLike[str], parent: Anthology) -> Venue: